Move ssl example models to their own file, rename to reflect they are for cifar and update references throughout. Update APFL client metrics
