+++ tests/smoke_tests/apfl_client_metrics.json
@@ -2,57 +2,57 @@
     "rounds": {
         "1": {
             "fit_metrics": {
-                "train - personal - accuracy": 0.6828,
-                "train - global - accuracy": 0.7109,
-                "train - local - accuracy": 0.55626
+                "train - personal - accuracy": 0.6078,
+                "train - global - accuracy": 0.6265,
+                "train - local - accuracy": 0.4890
             },
             "loss_dict": {
-                "backward": 1.2641,
-                "global": 1.1124,
-                "local": 1.4938
+                "backward": 1.4258,
+                "global": 1.2897,
+                "local": 1.6243
             },
             "evaluate_metrics": {
-                "val - personal - accuracy": 0.7093,
-                "val - global - accuracy": 0.6921,
-                "val - local - accuracy": 0.7168
+                "val - personal - accuracy": 0.6254,
+                "val - global - accuracy": 0.6757,
+                "val - local - accuracy": 0.5723
             },
-            "loss": 1.4237
+            "loss": 1.4845
         },
         "2": {
             "fit_metrics": {
-                "train - personal - accuracy": 0.8328,
-                "train - global - accuracy": 0.8203,
-                "train - local - accuracy": 0.7312
+                "train - personal - accuracy": 0.7845,
+                "train - global - accuracy": 0.7921,
+                "train - local - accuracy": 0.775
             },
             "loss_dict": {
-                "backward": 0.5003,
-                "global": 0.6467,
-                "local": 0.8632
+                "backward": 0.6868,
+                "global": 0.6557,
+                "local": 0.7524
             },
             "evaluate_metrics": {
-                "val - personal - accuracy": 0.8327,
-                "val - global - accuracy": 0.7825,
-                "val - local - accuracy": 0.7156
+                "val - personal - accuracy": 0.6237,
+                "val - global - accuracy": 0.6583,
+                "val - local - accuracy": 0.5203
             },
-            "loss": 0.5962
+            "loss": 1.1075
         },
         "3": {
             "fit_metrics": {
-                "train - personal - accuracy": 0.9078,
-                "train - global - accuracy": 0.8375,
-                "train - local - accuracy": 0.8875
+                "train - personal - accuracy": 0.8015,
+                "train - global - accuracy": 0.8078,
+                "train - local - accuracy": 0.78125
             },
             "loss_dict": {
-                "backward": 0.3403,
-                "global": 0.5057,
-                "local": 0.4182
+                "backward": 0.5872,
+                "global": 0.5809,
+                "local": 0.6471
             },
             "evaluate_metrics": {
-                "val - personal - accuracy": 0.8195,
-                "val - global - accuracy": 0.77,
-                "val - local - accuracy": 0.739
+                "val - personal - accuracy": 0.7182,
+                "val - global - accuracy": 0.7571,
+                "val - local - accuracy": 0.5864
             },
-            "loss": 0.5230
+            "loss": 0.8714
         }
     }
 }
